move invocation of verifier to GeneratorBase
